requirements.txt
# Core dependencies
matplotlib
mmcls>=0.20.1
mmcv-full>=1.4.4,<1.7.0
numpy
packaging
prettytable
scipy
torch
torchvision
timm
einops
opencv-python

# Optional dependencies
cityscapesscripts

# Test dependencies
codecov
flake8
interrogate
pytest
xdoctest>=0.10.0
yapf
